domain logic for creating linguistic variable using initial values via LinguisticVariableFactory.java

+++ src/main/java/ua/kpi/cad/linguisticvar/domain/Interval.java
@@ -8,6 +8,6 @@ import lombok.ToString;
 @Getter
 @ToString
 public class Interval {
-    private float leftBoundary;
-    private float rightBoundary;
+    private double leftBoundary;
+    private double rightBoundary;
 }
